polymarket-bot/requirements.txt
# --- Polymarket ---
py-clob-client==0.17.5

# --- Blockchain ---
web3>=7.0.0,<8.0.0

# --- Data & Analysis ---
scikit-learn==1.5.2
pandas==2.2.3
numpy==1.26.4

# --- Infrastructure ---
celery[redis]==5.4.0
redis==5.1.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.13.3

# --- API & bots ---
fastapi==0.115.4
uvicorn[standard]==0.32.0
python-telegram-bot[ext]==21.6
httpx==0.27.2
websockets==13.1

# --- Utilities ---
pydantic==2.9.2
pydantic-settings==2.6.1
loguru==0.7.2
python-dotenv==1.0.1
tenacity==9.0.0

# --- Tests ---
pytest==8.3.3
pytest-asyncio==0.24.0
